+++ CHANGELOG.md
@@ -28,6 +28,7 @@ All notable changes to this project will be documented in this file.
 - Forge and Befuddle now has a faction linked to it
 - Linked Forge and Befuddle
 - Event Reminder localization of time left should now be working correctly
+- Event Reminder localization of end time when using a custom format containing %a, %A, %b or %B should now be working correctly
 
 ### Wrath Classic
 - Bumped version